Backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-generativeai==0.5.4
pydantic==2.7.1
python-dotenv==1.0.1
pypdf==4.2.0
alembic==1.13.1
